feat(lectures): show chapter and card counts on the lecture card
Add a "N chapters · N cards" meta row to the lecture card, with counts
aggregated server-side in the existing getLectures list path.

- packages/api: add LectureListItem (Lecture + chapterCount/questionCount);
  LectureRepository.getAll now returns LectureListItem[], leaving the
  single-lecture paths on the base Lecture type
- server: move the repository factories into a side-effect-free
  repositories.ts taking an injected Database; getAll computes both counts
  in one grouped aggregate using COUNT(DISTINCT c.id) + LEFT JOINs to avoid
  the cartesian double-count and to return 0/0 for empty lectures
- web: render the meta row on LectureCard (always shown, incl. 0/0) with
  localized copy in en/de
- tests: unit-test the aggregation (counts + empty case); update the
  lecture-overview e2e mocks and assert the meta line renders
