+++ apps/server/src/repositories.ts
@@ -1,0 +1,276 @@
+import { type Database } from 'better-sqlite3';
+
+import {
+  type Chapter,
+  type ChapterRepository,
+  type Lecture,
+  type LectureListItem,
+  type LectureRepository,
+  type Question,
+  type QuestionRepository,
+} from '@athena/api';
+
+export function createLectureRepository(db: Database): LectureRepository {
+  return {
+    getAll: (): LectureListItem[] => {
+      return db
+        .prepare(
+          `SELECT
+             l.id, l.title, l.description,
+             COUNT(DISTINCT c.id) AS chapterCount,
+             COUNT(q.id)          AS questionCount
+           FROM lectures l
+           LEFT JOIN chapters c ON c.lectureId = l.id
+           LEFT JOIN questions q ON q.chapterId = c.id
+           GROUP BY l.id, l.title, l.description`,
+        )
+        .all() as LectureListItem[];
+    },
+    getById: (id: string): Lecture | undefined => {
+      return db.prepare('SELECT * FROM lectures WHERE id = ?').get(id) as
+        | Lecture
+        | undefined;
+    },
+    create: (lecture: Omit<Lecture, 'id'>): Lecture => {
+      const id = crypto.randomUUID();
+      db.prepare(
+        'INSERT INTO lectures (id, title, description) VALUES (?, ?, ?)',
+      ).run(id, lecture.title, lecture.description);
+      return { id, ...lecture };
+    },
+    update: (id: string, lecture: Omit<Lecture, 'id'>): Lecture | undefined => {
+      const result = db
+        .prepare('UPDATE lectures SET title = ?, description = ? WHERE id = ?')
+        .run(lecture.title, lecture.description, id);
+      if (result.changes === 0) return undefined;
+      return { id, ...lecture };
+    },
+    delete: (id: string): boolean => {
+      const result = db.prepare('DELETE FROM lectures WHERE id = ?').run(id);
+      return result.changes > 0;
+    },
+  };
+}
+
+export function createChapterRepository(db: Database): ChapterRepository {
+  return {
+    getById: (id: string): Chapter | undefined => {
+      return db.prepare('SELECT * FROM chapters WHERE id = ?').get(id) as
+        | Chapter
+        | undefined;
+    },
+    getByLectureId: (lectureId: string): Chapter[] => {
+      return db
+        .prepare('SELECT * FROM chapters WHERE lectureId = ? ORDER BY "order"')
+        .all(lectureId) as Chapter[];
+    },
+    getDistinctAssociations: (): string[] => {
+      const rows = db
+        .prepare(
+          "SELECT DISTINCT association FROM chapters WHERE association != '' ORDER BY association",
+        )
+        .all() as { association: string }[];
+      return rows.map((row) => row.association);
+    },
+    search: (query: string): (Chapter & { firstQuestion?: Question })[] => {
+      if (!query.trim()) return [];
+
+      const tokens = query.toLowerCase().split(/\s+/).filter(Boolean);
+      if (tokens.length === 0) return [];
+
+      // Search in questions (question text) and chapters (association)
+      // Use aliases for the main query
+      const questionConditions = tokens
+        .map(() => 'LOWER(q.question) LIKE ?')
+        .join(' AND ');
+
+      const associationConditions = tokens
+        .map(() => 'LOWER(c.association) LIKE ?')
+        .join(' AND ');
+
+      const questionParams = tokens.map((token) => `%${token}%`);
+      const associationParams = tokens.map((token) => `%${token}%`);
+
+      // Get chapters that match via ANY question or association
+      const chapters = db
+        .prepare(
+          `
+          SELECT DISTINCT c.* FROM chapters c
+          LEFT JOIN questions q ON q.chapterId = c.id
+          WHERE (${questionConditions}) OR (${associationConditions})
+          ORDER BY c."order"
+        `,
+        )
+        .all(...questionParams, ...associationParams) as Chapter[];
+
+      // Conditions for finding the matching question within a specific chapter
+      const innerQuestionConditions = tokens
+        .map(() => 'LOWER(question) LIKE ?')
+        .join(' AND ');
+
+      // Attach the best matching question to each chapter
+      return chapters.map((chapter) => {
+        // First try to find a question that matches the search query
+        let firstQuestion = db
+          .prepare(
+            `SELECT * FROM questions WHERE chapterId = ? AND (${innerQuestionConditions}) LIMIT 1`,
+          )
+          .get(chapter.id, ...questionParams) as Question | undefined;
+
+        // If no question matches (match was on association), fall back to the actual first question
+        if (!firstQuestion) {
+          firstQuestion = db
+            .prepare(
+              'SELECT * FROM questions WHERE chapterId = ? ORDER BY "order" LIMIT 1',
+            )
+            .get(chapter.id) as Question | undefined;
+        }
+
+        return { ...chapter, firstQuestion };
+      });
+    },
+    create: (chapter: Omit<Chapter, 'id'>): Chapter => {
+      const id = crypto.randomUUID();
+      db.prepare(
+        'INSERT INTO chapters (id, lectureId, association, "order") VALUES (?, ?, ?, ?)',
+      ).run(id, chapter.lectureId, chapter.association, chapter.order);
+      return { id, ...chapter };
+    },
+    update: (
+      id: string,
+      chapter: Partial<Omit<Chapter, 'id'>>,
+    ): Chapter | undefined => {
+      const existing = db
+        .prepare('SELECT * FROM chapters WHERE id = ?')
+        .get(id) as Chapter | undefined;
+      if (!existing) return undefined;
+      const updated = { ...existing, ...chapter };
+      db.prepare(
+        'UPDATE chapters SET lectureId = ?, association = ?, "order" = ? WHERE id = ?',
+      ).run(updated.lectureId, updated.association, updated.order, id);
+      return updated;
+    },
+    delete: (id: string): boolean => {
+      const result = db.prepare('DELETE FROM chapters WHERE id = ?').run(id);
+      return result.changes > 0;
+    },
+  };
+}
+
+export function createQuestionRepository(db: Database): QuestionRepository {
+  return {
+    getByChapterId: (chapterId: string): Question[] => {
+      return db
+        .prepare('SELECT * FROM questions WHERE chapterId = ? ORDER BY "order"')
+        .all(chapterId) as Question[];
+    },
+    getFirstByChapterId: (chapterId: string): Question | undefined => {
+      return db
+        .prepare(
+          'SELECT * FROM questions WHERE chapterId = ? ORDER BY "order" LIMIT 1',
+        )
+        .get(chapterId) as Question | undefined;
+    },
+    getFirstByLectureId: (lectureId: string): Record<string, Question> => {
+      // Get all chapters for this lecture, then get first question for each
+      const rows = db
+        .prepare(
+          `SELECT q.* FROM questions q
+           INNER JOIN (
+             SELECT chapterId, MIN("order") as minOrder
+             FROM questions
+             WHERE chapterId IN (SELECT id FROM chapters WHERE lectureId = ?)
+             GROUP BY chapterId
+           ) first ON q.chapterId = first.chapterId AND q."order" = first.minOrder`,
+        )
+        .all(lectureId) as Question[];
+
+      const result: Record<string, Question> = {};
+      for (const question of rows) {
+        result[question.chapterId] = question;
+      }
+      return result;
+    },
+    getAnnotatedChapterIdsByLecture: (lectureId: string): string[] => {
+      const rows = db
+        .prepare(
+          `SELECT DISTINCT q.chapterId FROM questions q
+           INNER JOIN chapters c ON q.chapterId = c.id
+           WHERE c.lectureId = ? AND q.isAnnotated = 1`,
+        )
+        .all(lectureId) as { chapterId: string }[];
+      return rows.map((row) => row.chapterId);
+    },
+    create: (question: Omit<Question, 'id'>): Question => {
+      const id = crypto.randomUUID();
+      db.prepare(
+        'INSERT INTO questions (id, chapterId, question, answer, "order", isAnnotated) VALUES (?, ?, ?, ?, ?, ?)',
+      ).run(
+        id,
+        question.chapterId,
+        question.question,
+        question.answer,
+        question.order,
+        question.isAnnotated ? 1 : 0,
+      );
+      return { id, ...question };
+    },
+    update: (
+      id: string,
+      question: Partial<Omit<Question, 'id'>>,
+    ): Question | undefined => {
+      const existing = db
+        .prepare('SELECT * FROM questions WHERE id = ?')
+        .get(id) as Question | undefined;
+      if (!existing) return undefined;
+      const updated = { ...existing, ...question };
+      db.prepare(
+        'UPDATE questions SET question = ?, answer = ?, "order" = ?, isAnnotated = ? WHERE id = ?',
+      ).run(
+        updated.question,
+        updated.answer,
+        updated.order,
+        // Ensure boolean/number compatibility
+        question.isAnnotated !== undefined
+          ? question.isAnnotated
+            ? 1
+            : 0
+          : existing.isAnnotated,
+        id,
+      );
+      return updated;
+    },
+    delete: (id: string): boolean => {
+      const result = db.prepare('DELETE FROM questions WHERE id = ?').run(id);
+      return result.changes > 0;
+    },
+    getQuestionCountsByLecture: (lectureId: string): number => {
+      const result = db
+        .prepare(
+          `SELECT COUNT(*) as count FROM questions q
+           INNER JOIN chapters c ON q.chapterId = c.id
+           WHERE c.lectureId = ?`,
+        )
+        .get(lectureId) as { count: number };
+      return result.count;
+    },
+    getQuestionCountsPerChapter: (
+      lectureId: string,
+    ): Record<string, number> => {
+      const rows = db
+        .prepare(
+          `SELECT c.id as chapterId, COUNT(q.id) as count
+           FROM chapters c
+           LEFT JOIN questions q ON q.chapterId = c.id
+           WHERE c.lectureId = ?
+           GROUP BY c.id`,
+        )
+        .all(lectureId) as { chapterId: string; count: number }[];
+      const result: Record<string, number> = {};
+      for (const row of rows) {
+        result[row.chapterId] = row.count;
+      }
+      return result;
+    },
+  };
+}
